pyproject.toml
[tool.poetry]
name = "imagesecrets-api"
version = "0.5.0"
description = "A tool for encoding and decoding messages into/from images."
authors = ["Lukas Kucera <lukas.kucera.g@protonmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://gitlab.com/Lkms19/imagesecrets-api"

[tool.black]
line-length = 79
include = '^.+\.py$'

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["image_secrets", "tests"]

[tool.flake8]
max-line-length = 79
extend-ignore = "E501,"
max-complexity = 10

[tool.pytest.ini_options]
filterwarnings = 'ignore::DeprecationWarning:asynctest.*:'
markers = [
    "disable_autouse: mark to disable autouse fixtures for the test",
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
fastapi-login = "^1.6.0"
fastapi-mail = "^0.3.7"
pydantic = {extras = ["email", "ujson"], version = "^1.8.2"}
python-multipart = "^0.0.5"
aiofiles = "^0.7.0"
requests = "^2.25.1"
uvicorn = "^0.13.4"
tortoise-orm = { extras = ["asyncpg"], version = "^0.17.2" }
numpy = "^1.20.2"
Pillow = "^8.2.0"
bcrypt = "^3.2.0"
python-dotenv = "^0.17.1"
python-magic = "^0.4.14"
Jinja2 = "^3.0.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
pytest = "^6.2.3"
pytest-mock = "^3.6.1"
pytest-cov = "^2.12.0"
pytest-asyncio = "^0.15.1"
asynctest = "^0.13.0"
devtools = "^0.6.1"
mypy = "^0.812"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
